agents/Manager/requirements.txt
google-adk
deprecated
fastapi
uvicorn[standard]
python-dotenv
